Refactor types from internal, improves public types for mixins
