+++ packages/@ember/-internals/runtime/index.d.ts
@@ -1,37 +1,14 @@
-import { EmberClassConstructor, Objectify } from './types/-private/types';
-import PublicCoreObject from './types/core';
-import PublicEmberObject from './types/index';
-import Evented from './types/evented';
+import CoreObject from './lib/system/core_object';
+import Object from './lib/system/object';
+import Evented from './lib/mixins/evented';
 export { default as copy } from './lib/copy';
 
 export const TargetActionSupport: any;
 export function isArray(arr: any): boolean;
 
-// The public version doesn't export some deprecated methods.
-// However, these are still used internally. Returning `any` is
-// very loose, but it's essentially what was here before.
-export class CoreObject extends PublicCoreObject {
-  static extend<Statics, Instance>(
-    this: Statics & EmberClassConstructor<Instance>,
-    ...args: any[]
-  ): Objectify<Statics> & EmberClassConstructor<Instance>;
-  static reopen(...args: any[]): any;
-  static reopenClass(...args: any[]): any;
-}
+export { CoreObject, Object };
 
 export const FrameworkObject: any;
-
-// The public version doesn't export some deprecated methods.
-// However, these are still used internally. Returning `any` is
-// very loose, but it's essentially what was here before.
-export class Object extends PublicEmberObject {
-  static extend<Statics, Instance>(
-    this: Statics & EmberClassConstructor<Instance>,
-    ...args: any[]
-  ): Objectify<Statics> & EmberClassConstructor<Instance>;
-  static reopen(...args: any[]): any;
-  static reopenClass(...args: any[]): any;
-}
 
 export function _contentFor(proxy: any): any;
 
@@ -39,3 +16,5 @@ export const A: any;
 export const ActionHandler: any;
 export { Evented };
 export function typeOf(obj: any): string;
+
+export { default as Observable } from './lib/mixins/observable';
