+++ packages/@ember/object/type-tests/evented/index.test.ts
@@ -1,7 +1,7 @@
 import { expectTypeOf } from 'expect-type';
 
 import Mixin from '@ember/object/mixin';
-import Observable from '@ember/object/observable';
+import Evented from '@ember/object/evented';
 
 // A very naive test that at least makes sure we can import this
-expectTypeOf<typeof Observable>().toMatchTypeOf<Mixin>();
+expectTypeOf<typeof Evented>().toMatchTypeOf<Mixin>();
