+++ packages/@ember/object/type-tests/mixin/index.test.ts
@@ -1,0 +1,9 @@
+import { expectTypeOf } from 'expect-type';
+
+import Mixin from '@ember/object/mixin';
+
+let newMixin = Mixin.create({
+  foo: 'bar',
+});
+
+expectTypeOf(newMixin).toMatchTypeOf<Mixin>();
