--- packages/@ember/object/types/core.d.ts
@@ -1,1 +1,0 @@
-export { default } from '../../-internals/runtime/types/core';
