--- packages/@ember/object/types/observable.d.ts
@@ -1,2 +1,0 @@
-import Observable from '@ember/-internals/runtime/types/observable';
-export default Observable;
